# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WWM Companion Local API
  version: 0.1.0
servers:
  - url: http://127.0.0.1:3200/api/v1
paths:
  /health:
    get:
      responses:
        "200":
          description: Local service health
  /regions:
    get:
      responses:
        "200":
          description: Available regions
  /map-tiles/{map}/{language}/{zoom}/{tile}:
    get:
      parameters:
        - { in: path, name: map, required: true, schema: { type: string } }
        - { in: path, name: language, required: true, schema: { type: string } }
        - { in: path, name: zoom, required: true, schema: { type: integer } }
        - { in: path, name: tile, required: true, schema: { type: string } }
      responses:
        "200":
          description: Proxied official raster tile
  /regions/{id}/progress:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Region and category completion totals
  /regions/{id}/source-categories:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full official category taxonomy and POI totals for a region
  /pois:
    get:
      parameters:
        - in: query
          name: region_id
          schema: { type: string }
        - in: query
          name: categories
          schema: { type: string }
        - in: query
          name: source_categories
          schema: { type: string }
        - in: query
          name: missing_only
          schema: { type: boolean }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: floor
          schema: { type: string }
      responses:
        "200":
          description: Filtered POIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/POI"
  /pois/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: POI details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POI"
        "404":
          description: POI not found
  /pois/{id}/complete:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated POI
  /pois/{id}/uncomplete:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated POI
  /screenshots/analyze:
    post:
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, poi_id]
              properties:
                image: { type: string, format: binary }
                poi_id: { type: string }
                optional_note: { type: string }
      responses:
        "200":
          description: Validated analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
  /routes:
    post:
      responses:
        "200":
          description: Deterministic missing-POI route
  /settings:
    get:
      responses:
        "200":
          description: Local settings
    put:
      responses:
        "200":
          description: Updated settings
  /progress/export:
    get:
      responses:
        "200":
          description: Portable progress JSON
  /progress/import:
    post:
      responses:
        "200":
          description: Import count
components:
  schemas:
    ReferenceImage:
      type: object
      required: [url, alt, source_url]
      properties:
        url: { type: string, format: uri }
        alt: { type: string }
        caption: { type: string }
        source_url: { type: string, format: uri }
    Provenance:
      type: object
      required: [source, source_type, verification_status, patch_version]
      properties:
        source: { type: string }
        source_url: { type: string, format: uri }
        source_type: { type: string, enum: [dev_fixture, manual, licensed, official, community_sync] }
        verification_status: { type: string, enum: [fixture, unverified, verified, source_synced] }
        patch_version: { type: string }
        last_verified_date: { type: string, format: date }
    POI:
      type: object
      required:
        - id
        - name
        - category
        - region_id
        - coordinates
        - requirements
        - navigation_steps
        - solution_steps
        - reference_images
        - status
        - provenance
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string, enum: [chest, oddity, puzzle, quest] }
        region_id: { type: string }
        subregion: { type: string }
        coordinates:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: number }
        floor: { type: string }
        entrance: { type: string }
        nearest_landmark: { type: string }
        requirements:
          type: array
          items: { type: string }
        navigation_steps:
          type: array
          items: { type: string }
        solution_steps:
          type: array
          items: { type: string }
        common_mistake: { type: string }
        reference_images:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceImage"
        guide_provenance:
          $ref: "#/components/schemas/Provenance"
        status:
          type: string
          enum: [completed, not_completed, uncertain]
        source_category: { type: string }
        provenance:
          $ref: "#/components/schemas/Provenance"
    AnalysisResult:
      type: object
      required: [analysis_id, poi_id, status, confidence, evidence, cached, provider]
      properties:
        analysis_id: { type: string }
        poi_id: { type: string }
        status:
          type: string
          enum: [completed, not_completed, uncertain]
        confidence: { type: number, minimum: 0, maximum: 1 }
        evidence:
          type: array
          items: { type: string }
        cached: { type: boolean }
        provider:
          type: string
          description: Provider and resolved model identifier